# docs/open-api/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "DiffScope Synthesis Platform API",
    "version": "0.0.0",
    "description": "DSSP HTTP API surface represented by this service. A server implementation can support a subset of the broader DSSP API."
  },
  "servers": [
    {
      "url": "http://localhost:13711",
      "description": "Default local server URL. The actual host and port are configured by the service."
    }
  ],
...
  "paths": {
...
    "/api/arch": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getArchitectureList",
        "summary": "List architectures",
        "description": "Returns metadata for every registered architecture.",
        "parameters": [
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ArchitectureList"
          },
          "404": {
            "$ref": "#/components/responses/ArchitectureNotFound"
          },
          "422": {
            "$ref": "#/components/responses/ArchitectureOperationError"
          }
        }
      }
    },
    "/api/arch/{arch_id}": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getArchitecture",
        "summary": "Get architecture metadata",
        "description": "Returns metadata for one registered architecture.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArchID"
          },
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Architecture"
          },
          "404": {
            "$ref": "#/components/responses/ArchitectureNotFound"
          },
          "422": {
            "$ref": "#/components/responses/ArchitectureOperationError"
          }
        }
      }
    },
...
    "/api/synth/duration": {
      "post": {
        "tags": [
          "Synthesis"
        ],
        "operationId": "postSynthDuration",
        "summary": "Generate phoneme timing",
        "description": "Generates phoneme timing for duration notes. Set stream to true to receive newline-delimited JSON events.",
        "requestBody": {
          "$ref": "#/components/requestBodies/DurationRequest"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/DurationResult"
          },
          "204": {
            "$ref": "#/components/responses/EnvTagMatched"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "422": {
            "$ref": "#/components/responses/DurationOperationError"
          }
        }
      }
    },
...
    "/api/synth/audio": {
      "post": {
        "tags": [
          "Synthesis"
        ],
        "operationId": "postSynthAudio",
        "summary": "Generate audio",
        "description": "Generates encoded audio as a data URL. Set stream to true to receive newline-delimited JSON events.",
        "requestBody": {
          "$ref": "#/components/requestBodies/AudioRequest"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AudioResult"
          },
          "204": {
            "$ref": "#/components/responses/EnvTagMatched"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "422": {
            "$ref": "#/components/responses/AudioOperationError"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "ArchID": {
        "name": "arch_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Architecture identifier."
      },
...
      "DisplayLanguage": {
        "name": "display_language",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Preferred language for display metadata."
      }
    },
    "requestBodies": {
...
      "DurationRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DurationRequest"
            }
          }
        }
      },
...
      "AudioRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AudioRequest"
            }
          }
        }
      }
    },
    "responses": {
      "ArchitectureList": {
        "description": "Architecture metadata from every registered architecture.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArchitectureMetadataList"
            }
          }
        }
      },
      "Architecture": {
        "description": "Architecture metadata.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArchitectureMetadata"
            }
          }
        }
      },
      "ArchitectureNotFound": {
        "description": "The requested architecture does not exist.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArchitectureNotFoundAPIError"
            }
          }
        }
      },
      "ArchitectureOperationError": {
        "description": "The architecture metadata request was syntactically valid but could not be completed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArchitectureOperationAPIError"
            }
          }
        }
      },
...
      "DurationResult": {
        "description": "Duration generation result. Non-streaming requests return a complete JSON response. Streaming requests return newline-delimited JSON events; an error event can appear after the HTTP status has already been sent.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DurationResponse"
            }
          },
          "application/x-ndjson": {
            "itemSchema": {
              "$ref": "#/components/schemas/DurationStreamEvent"
            }
          }
        }
      },
...
      "AudioResult": {
        "description": "Audio generation result. Non-streaming requests return a complete JSON response. Streaming requests return newline-delimited JSON events; an error event can appear after the HTTP status has already been sent.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AudioResponse"
            }
          },
          "application/x-ndjson": {
            "itemSchema": {
              "$ref": "#/components/schemas/AudioStreamEvent"
            }
          }
        }
      },
      "EnvTagMatched": {
        "description": "The supplied environment tag matches the current environment tag, so no response body is returned."
      },
      "BadRequest": {
        "description": "The request body is not valid JSON, contains more than one JSON value, or fails request validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BadRequestError"
            }
          }
        }
      },
...
      "DurationOperationError": {
        "description": "The duration request was syntactically valid but could not be completed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DurationAPIError"
            }
          }
        }
      },
...
      "AudioOperationError": {
        "description": "The audio request was syntactically valid but could not be completed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AudioAPIError"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ArchitectureMetadataList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ArchitectureMetadata"
        }
      },
      "ArchitectureMetadata": {
        "type": "object",
        "required": [
          "id",
          "name",
          "pronunciation_mode",
          "phoneme_mode",
          "parameters",
          "audio_dependencies"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Architecture identifier."
          },
          "name": {
            "type": "string"
          },
          "pronunciation_mode": {
            "$ref": "#/components/schemas/PronunciationMode"
          },
          "phoneme_mode": {
            "$ref": "#/components/schemas/PhonemeMode"
          },
          "parameters": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ArchitectureParameterMetadata"
            }
          },
          "audio_dependencies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "PronunciationMode": {
        "type": "string",
        "enum": [
          "FULL",
          "SKIP"
        ]
      },
      "PhonemeMode": {
        "type": "string",
        "enum": [
          "FULL",
          "TOKEN_ONLY",
          "SKIP"
        ]
      },
      "ArchitectureParameterMetadata": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/DirectArchitectureParameterMetadata"
          },
          {
            "$ref": "#/components/schemas/IndirectArchitectureParameterMetadata"
          }
        ]
      },
...
      "DurationRequest": {
        "type": "object",
        "required": [
          "context",
          "input"
        ],
        "properties": {
          "context": {
            "$ref": "#/components/schemas/MultiSingerContext"
          },
          "input": {
            "$ref": "#/components/schemas/DurationInput"
          },
          "env_tag": {
            "$ref": "#/components/schemas/EnvTag"
          },
          "stream": {
            "type": "boolean",
            "description": "When true, the response uses application/x-ndjson and emits progress events."
          }
        },
        "additionalProperties": false
      },
...
      "AudioRequest": {
        "type": "object",
        "required": [
          "context",
          "input"
        ],
        "properties": {
          "context": {
            "$ref": "#/components/schemas/MultiSingerContext"
          },
          "input": {
            "$ref": "#/components/schemas/AudioInput"
          },
          "env_tag": {
            "$ref": "#/components/schemas/EnvTag"
          },
          "stream": {
            "type": "boolean",
            "description": "When true, the response uses application/x-ndjson and emits progress events."
          }
        },
        "additionalProperties": false
      },
...
      "MultiSingerContext": {
        "type": "object",
        "required": [
          "arch",
          "arch_extra",
          "singers"
        ],
        "properties": {
          "arch": {
            "type": "string",
            "description": "Architecture identifier."
          },
          "arch_extra": {
            "description": "Architecture-specific JSON payload."
          },
          "singers": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/Singer"
            }
          }
        },
        "additionalProperties": false
      },
...
      "EnvTag": {
        "type": "string",
        "description": "Previously returned environment tag. If it matches the current environment tag, the service can return 204."
      },
...
      "DurationResponse": {
        "type": "object",
        "required": [
          "state",
          "output"
        ],
        "properties": {
          "state": {
            "const": "COMPLETE"
          },
          "output": {
            "$ref": "#/components/schemas/DurationOutput"
          },
          "env_tag": {
            "$ref": "#/components/schemas/ResponseEnvTag"
          }
        },
        "additionalProperties": false
      },
...
      "AudioResponse": {
        "type": "object",
        "required": [
          "state",
          "output"
        ],
        "properties": {
          "state": {
            "const": "COMPLETE"
          },
          "output": {
            "$ref": "#/components/schemas/AudioOutput"
          },
          "env_tag": {
            "$ref": "#/components/schemas/ResponseEnvTag"
          }
        },
        "additionalProperties": false
      },
...
      "DurationStreamEvent": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/StreamStateEvent"
          },
          {
            "$ref": "#/components/schemas/DurationResponse"
          },
          {
            "$ref": "#/components/schemas/DurationAPIError"
          }
        ]
      },
...
      "AudioStreamEvent": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/StreamStateEvent"
          },
          {
            "$ref": "#/components/schemas/AudioResponse"
          },
          {
            "$ref": "#/components/schemas/AudioAPIError"
          }
        ]
      },
...
      "ArchitectureNotFoundAPIError": {
        "type": "object",
        "required": [
          "state",
          "code",
          "message"
        ],
        "properties": {
          "state": {
            "const": "ERROR"
          },
          "code": {
            "$ref": "#/components/schemas/ArchitectureNotFoundErrorCode"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ArchitectureOperationAPIError": {
        "type": "object",
        "required": [
          "state",
          "code",
          "message"
        ],
        "properties": {
          "state": {
            "const": "ERROR"
          },
          "code": {
            "$ref": "#/components/schemas/ArchitectureOperationErrorCode"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "BadRequestError": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "DurationAPIError": {
        "type": "object",
        "required": [
          "state",
          "code",
          "message"
        ],
        "properties": {
          "state": {
            "const": "ERROR"
          },
          "code": {
            "$ref": "#/components/schemas/DurationErrorCode"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "AudioAPIError": {
        "type": "object",
        "required": [
          "state",
          "code",
          "message"
        ],
        "properties": {
          "state": {
            "const": "ERROR"
          },
          "code": {
            "$ref": "#/components/schemas/AudioErrorCode"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ArchitectureNotFoundErrorCode": {
        "type": "string",
        "enum": [
          "UNKNOWN_ARCH"
        ]
      },
      "ArchitectureOperationErrorCode": {
        "type": "string",
        "enum": [
          "INTERNAL_ERROR"
        ]
      },
...
    }
  }
}